Save Volatile FPRs in PicBuilder before C call
In PicBuilder, assembly helper to interface call contained a bug
where it was not preserving volatile Floating point registers before
calling a C function. In one of the workload, this was causing issue where
the C helper was clobbering the FPRs causing corrupted value in the register
which later manifests in to java exception in one of the unit tests. This commit
fixes that by storing and restoring FPRs before call.

Fixes: #9354, #9445

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>