Set Code Cache alignment to 16 bits on IBM Z
On IBM Z platform, to support load and store quadword instruction from
snippet, snippet address should be aligned to quadword boundary. In JIT
compilation this works as when emitting snippet, we align the snippets
according to its size. But in AOT load, there was a possibility that it
mess up with the alignment for start of snippet address. By setting the
code cache memory alignment to 16, we make sure, that in AOT load, the
snippet from where it is loading or storing, will be aligned.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>