Use map of symrefs in OSR liveness data to those in current trees
EscapeAnalysisTools relied upon OSR liveness data to supply arguments
to calls to eaEscapeHelper in OSR induce blocks.  This enables more stack
allocation of objects under OSR.  However, after various transformations
have been applied to the trees, the liveness data cannot be relied upon
directly.

Instead, Pre-EscapeAnalysis must request a DefiningMap from the
OSRCompilationData, which provides a map from the symbol references that
existed at the point of OSR liveness analysis to the definitions that
exist in the trees at the current point in the compilation.

For each of the relevant auto and pending push symrefs from the liveness
data at the induce block, EscapeAnalysisTools::processSymbolReferences
will add the defining symrefs mapped to by the original symref to the
call to eaEscapeHelper that it inserts.

The use of the DefiningMap can be disabled by setting the environment
variable TR_DisableEAEscapeHelperDefiningMap.

Co-authored-by:  Yi Zhang <yizhang@ca.ibm.com>

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>