Add function for aborting compilation with bad value type operations
Currently, the IL generator will abort compilation when an operation
with an unresolved CP reference that may need value type handling
is encountered. To aid readability and maintainability, this commit
adds a new helper function that encapsulates the code that does the
abort. All instances of similar code are replaced with a call
to this new function.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>