Don't free twice
If there is a failure while relocating a method, if the
_relocationStatus field in TR_RelocationRuntime is set to
RelocationFailure, during the cleanup addFreeBlock is called. However,
if the compiled copy of the exception table was used (which can happen
if disableDelayRelocationForAOTCompilations is specified on the
-Xjit/-Xaot option), addFreeBlock will also be called in
CompilationThread.cpp.

While currently there might be some implicit coupling that prevents the
double free from happening, this is a bug that's just waiting to show up
in the future.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>