Fix crash printing to verbose log
After an AOT Compilation, the compiler either delays the relocation of
the body, or the relocation can fail. When this occurs, the
J9JITExceptionTable that was allocated has to freed. As part of this
freeing process, information is printed out to the verbose log if the
code cache verbose options are set. Part of the information that is
printed out is the name of the method. This information is retrieved
from the J9JITExceptionTable. However, this information can be NULL
because it gets populated when the table is relocated.

This commit adds a NULL check, as well as some sanity checks to ensure
that those fields are only NULL in the expected scenarios.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>