Fix Character.isUpperCase inlining on POWER9
Previously, the code for inlining Character.isUpperCase in the Power
codegen for P9 chips had an out-of-range immediate in an li instruction.
As it turns out, this issue was benign since the cmprb instruction that
used this value would only read the bottom 16 bits of the register, but
this would still cause an assertion failure in the binary encoder. This
has now been fixed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>