Collect stats about sizes of different types of messages sent
I'm keeping track of the size of different types of messages sent using msg.type()
and I'm using the class TR_Stats to store avg,min,max and stddev of every type of message.
I declared a static array in CommunicationStream.hpp and I'm using that to store in.
issue: #9708

Signed-off-by: Eman Elsabban <eman.elsaban1@gmail.com>