Change offset/pointer calculation
The offsets for pointers into the ROMClass section of the SCC are
calculated using the start of the ROMClass section. The offsets for
pointers into the metadata section of the SCC are calculated using
the start of the metadata section (which grows backwards, towards the
ROMClass section). The offset is then left shifted one bit; the low
bit is used to determine whether the offset is calculated from the start
of the ROMClass section or the start of the metadatasection.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>