Use correct APIs in Validation Record
TR_RelocationRecordValidateStaticField inherits from
TR_RelocationRecordValidateInstanceField (which inherits from
TR_RelocationRecordValidateClass). Unlike ValidateInstanceField and
ValidateClass, ValidateStaticField stores an offset to the J9ROMClass of
the class of the field, rather than an offset to the class chain of the
class of the field.

This commit adds a virtual method to these classes to ensure that the
right API is called.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>