Update numActiveThreads at the proper location at the server
As described in #9762, `_numActiveThreads` is not
incremented when the thread is unloading the class
or doing CHTable update. It could cause a race
condition when another thread is timed out
on waiting for the first thread to complete,
the second thread that’s timed out will clear the
client session cache because `_numActiveThreads` is 0.

Fixes #9762, #9694

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>