Fix dependencies generated in instanceOf Evaluator
On IBM Z, for Non Compressed Ref configuration, instanceOf evaluator
uses a Store Quadword instruction to store the data into dynamic cache.
To store the data in OOL evaluator was using new register pair assigning
objClassReg and castClassReg as register pair which was causing an issue
in register assigner as on mainline sequence these registers were not
used as register pair. To fix this issue and minimizing pressure on
register assigner, this fix in this commit reuses the register pair used
in mainline sequence to load from cache instead of allocating new one.

Fixes: #9770

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>