Remove unnecessary masking of lockword offset
Previously, there was one location where the offset of the lockword used
in an li instruction was masked to include only the bottom 16 bits.
While this is currently correct AFAIK, since it can never be negative
and cannot currently exceed 0x7fff legitimately, it may cause the assert
on the range of the immediate to be silenced incorrectly should this
change in the future. In line with this, the lockword offset is now used
directly.

Signed-off-by: Ben Thomas <ben@benthomas.ca>