Fix Power ArrayCopyBNDCHK handling of negative constants
Previously, the Power codegen's evaluator for ArrayCopyBNDCHK implicitly
assumed that neither of its children was a negative constant by failing
to check them against LOWER_IMMED when generating cmpi/twi instructions.
This could result in incorrect behaviour if exactly one of the children
is a constant which happens to be less than LOWER_IMMED. Prior to the
binary encoder refactor, this would result in the immediate being
silently truncated to 16 bits. This can cause the check to behave
incorrectly and can cause the bound check to pass when it shouldn't.

This has now been corrected by adding the necessary checks of constant
values against LOWER_IMMED.

Signed-off-by: Ben Thomas <ben@benthomas.ca>