Remove a questionable assumption about the max arraylet size
Previously, the Power codegen assumed that the maximum size of an
arraylet leaf in elements could be fit into a 16-bit immediate field.
This may not always be true, so could lead to incorrect behaviour.
Handling for when the maximum number of elements is > 0x7fff has now
been added.

Signed-off-by: Ben Thomas <ben@benthomas.ca>