Immediate field handling cleanup on Power
buildDirectDispatch was incorrectly truncating an immediate value before
generating an lis instruction. This was corrected.

generatePrologue in PPCRecompilation.cpp also had the same problem and has
been corrected.

Added in several new asserts to give better information in cases where an out
of range value is attempted to be used as the immediate value for an
instruction.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>