Disable jitdump at the server
During jitdump, the server could try to read and write
messages to the client to retrieve information
such as TR_J9ServerVM::sampleSignature() could be
triggered as part of printIRTrees(). Since at the time
the server has encountered a crash, any stream related exception
won’t be handled at this point. Since at the moment there is
no value with the jitdmp at the server, we could disable it for now.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>