Check for assignment of NULL to array of value type
In the prototype implementation of vlaue types, if the component type of
an array is a value type, a NullPointerException must be thrown when a
null reference is assigned to an element of the array.  If the array's
component type might be a value type, add a run-time test of whether
the array actually is a value type guarding a NULLCHK of the value that
is being assigned to the array element.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>