Disable AutoSIMD when PIV is a child of a commoned node
AutoSIMD was generating incorrect IL due to the side effect of node
commoning. Upon investigating further it was clear that to come up with
a general solution, it would require time. Moreover, it appears that even
for correct IL, AutoSIMD was failing in functionality tests.
This suggest that there may be bug in the codegen as well. Thus, it was
decided that for such cases, AutoSIMD will be disabled until a generalized
solution is ready.

This kind of optimization is currently available only in z Systems. and
does not affect amd64 / powerpc.

For more details see #9446

Signed-off-by: Mohammad Nazmul Alam <mohammad.nazmul.alam@ibm.com>