Avoid gpProtect path when -Xrs is set
Threads are signal protected when created by the JVM or on
the first call-in to the VM.  Subsequent call-ins can see
that the thread is already gpProtected and avoid adding the
gpProtection again.

In the case that the user is running -Xrs, and the thread is
not yet gpProtected, we end up doing a couple of c calls
before we realize signal protection isn't required and then
call through the passed in function pointer.

We can check for -Xrs early and avoid a couple of c calls on
the first call-in.  This should be zero cost on the common
case where the thread is already gpProtected and minimal cost
on the uncommon case where a thread isn't gpProtected yet.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>