Trap when DF flag is incorrectly set
Check DF flag at method entry/exit, object/array new, arraycopy and
calls in mainline. Trap if it is incorrectly set.

The check is off by default, it can be turned on by setting
environment variable `TR_enableBreakOnDFSet`.

Incorrect DF flag is the cause of a memory corruption defect in which a
rep stosb instruction is executed in the wrong direction. This defect
is intermittent and hasn't been reproduced. This change will provide a
way to catch such case in our builds.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>