Prefetch resolved implementor methods for CHTable
`CollectImplementors` is used by CHTable to find
the number of implementors of a given method.
It works by recursively walking the class tree
and creating a resolved method for each implementor.
This results in many remote queries in JITServer.

This commit adds a new helper class - `CollectResolvedImplementors`.
Its purpose is to walk the class tree and pre-emptively cache
all resolved methods that might be created by `CollectImplementors`
in one message.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>