Fix deadlock during jitdump when compilation thread releases VM access
This change fixes a deadlock which can occur where a thread is trying
to acquire exclusive VM access but is unable to. The following
conditions can be observed:

- GC thread is attempting to acquire exclusive, but it is waiting for
`vm->exclusiveAccessResponseCount` to be zero
- All compilation threads except one are stuck waiting on the dump
monitor to trigger a dump agent
- One thread is executing the jitdump and it has just released VM
access and is waiting on the diagnostic thread to finish compiling the
queued JIT method on the compilation queue
- The diagnostic thread is attempting to process the compilation queue
but it attempts to acquire VM access, so it blocks because GC wants
exclusive

This three way deadlock can be resolved by allowing the GC to acquire
exclusive VM access. The reason it is unable to is because the jitdump
thread (crashed thread) has released VM access even though it was not
holding it in the first place. This decrements
`vm->exclusiveAccessResponseCount` one too many times and the GC thread
is unable to acquire exclusive access so that everyone can proceed.

The fix is to simply acquire VM access before requesting the jitdump
recompilation since we will go down a code path that assumes we have VM
access already. i.e. synchronous compilations are synonymous with
requesting the compile on an application thread, which should always
hold (non-exclusive) VM access.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>