Stored the ram classes before sending the compilationreq to server
Stored the ramClasses in an unordered_set before sending the compilation
request to the server. I check if the ram class (clazz) is cached in the set.
if it is not stored then it will be cached in the set and will send the corresponding
ROM by setting the serializeClass bool to true. If the clazz is already in the cache set
this means it has already been sent to the server. So send the clazz again but with empty ROM
class (empty string for ROM Class) by setting serializeClass bool to false.
issue: #9708

Signed-off-by: Eman Elsabban <eman.elsaban1@gmail.com>