Fix J9BuildFlags aliases
The DDR tools from OMR, normally generate constants with names as they
appear in the source code. Unfortunately, the legacy tools used in IBM
builds yield different names for 'build flags'. A field override was
added to provide compatibility with those legacy tools for s390
  J9VM_ARCH_S390=arch_s390
but not for other architectures. J9ConfigFlags was introduced in an
attempt to handle this difference in naming convention, but it
incorrectly assumed the fields it referenced would be dynamically
generated according to the blob found in the associated core file,
rather than being fixed at build time. The set of aliases here replaces
the ineffective mapping in J9ConfigFlags.
The result is that OpenJ9 builds have additional fields in J9BuildFlags
that are copied from the OMR-derived names. For example, arch_x86 is
present when the legacy tools are used; when OMR tools are used,
arch_x86 is a copy of J9VM_ARCH_X86.

Remove J9ConfigFlags and redirect references to J9BuildFlags.

[ci skip]

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>