A script to parse the perf tool result
When we look at community raised performance issue on OpenJ9, we sometimes need to ask user to collect profile information. The output from the perf tool, i.e. perf.data is kernel depended. So we will need the user to process the data on their side before sending the raw data to us. This tool will parse the raw profile data and produce a text file which list the metheds by number of ticks.

Assists to close: #8353
Signed-off-by: Guangji Beverly Xu <guangjxu@ca.ibm.com>