Use _NSGetEnviron() on OSX to retrieve environ
Direct access to `environ` on OSX should use _NSGetEnviron(), as
`environ` is only directly accessible via that environment
subroutine from shared libraries and bundles.

---

`environ` is a list of strings that is made available when a
process begins. Each string has the convention "name=value" which
is set in the user environment.

`environ` can be directly accessed, except from shared libraries
and bundles, since it is only available to the loader `ld` when a
complete program is being linked. The environment routines
`getenv`, `setenv` and `putenv` can be used to access `environ`
whether referenced by the loader or in shared libraries and
bundles. If direct access to `environ` is needed in shared
libraries or bundles for OSX, `_NSGetEnviron` (defined in
`crt_externs.h`) can be used to retrieve `environ` at runtime.

Source: https://opensource.apple.com/source/Libc/Libc-1353.41.1/man/FreeBSD/environ.7.auto.html

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>