Use Persistent CH Table for CCV Caching
Prior to this commit, if Class Chain Validation Caching was enabled, the
result was cached using the offset of the ROM Class of the RAM Class to
be validated. However, this is not functionally correct. It is possible
for two J9Classes to have completely different class hierarchies and yet
have the same ROM Class; that is, even though two J9Classes have the
same J9ROMClass, one or more of their super classes may not have the
same J9ROMClass as the other. Thus, the class chains of these two
classes would be completely different. If the result of the class chain
validation is cached using the J9ROMClass, the validation may
incorrectly succeed.

Additionally, if a class gets redefined for any reason, the cached CCV
result for it AND all of its subclasses must be reset.

For these two reasons, the CCV Result is cached in the Persistent CH
Table. This allows the result to be keyed against the J9Class, and also
provides the means to reset the cached value should the hierarchy
change. The consequence of doing so, however, is that CCV Caching is now
dependant on the CHTable not being disabled.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>