Group cache field and static info in JITServer
- `TR_J9ServerVM::jitFieldsAreSame`/`jitStaticsAreSame` are mostly called
during ILGen for checking if the symbol reference for a field
has already been created. We iterate through all existing symbol
references of the corresponding type and determine if any existing
symref describes the current field.
This commit optimizes the process by prefetching info for all
fields and statics in a method in one message.

- Implement caching for `jitStaticsAreSame`
and merges message hanlding code for fields and statics into one,
since they are funcionally identical.

- Simplify client-side code for finding field data.
non-JITServer code doesn't look for field data if both fields are
resolved and their value offsets are different. This is an optimization
that shouldn't affect functional correctness, so this commit always
looks for field data, because it improves caching and simplifies the
code.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>