Inlining POWER10 arrayCopy
Deleting inlineVSXArrayCopy: this was a piece of work with experimental characteristics
it bloated the codegen to such an extent that it hurt general workload performance 5-10%
range. It was disabled many years ago. I took this opportunity to delete it completely.

Adding the POWER10 inlining of arraycopy: this version is only applicable to Concurrent
Scavenge. We are limited somewhat in using registers.

Adding the capability to call the POWER10 arraycopy helpers too.

Deleting LEArrayCopy related stuffs.

For performance reason, adjusting straight inlining to 64-byte.

For performance reason, using lxv/stxv instead

For performance reason, special-case for residue in 1/2/4/8 bytes

Adding a conditional on byteOffset being non-zero

Rebase to resolve conflicts

Fix a typo in comment.

Signed-off-by: Julian Wang <zlwang@ca.ibm.com>