JEP 378: Text Blocks graduates in JDK 15
The Text Blocks feature has left preview so the new functions
added by it are no long deprecated.

I've left the inclusion of the functions in JDK 13+ preprocessor
tags so as not to break any JDK 13+ testing as after the next
feature release - 0.22.0 in Sept - neither 13 or 14 will be in
support.

It doesn't seem worthwhile to add !15 specific preprocessor tags
around the @deprecated annotations.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>