Remove unnecessary monitor enter during VM exit
The nativeLibraryMonitor may be owned by a thread which is blocked by
exclusive VM access (acquired during VM exit), which results in a
deadlock on exit.

It is not necessary to enter the monitor in order to block native
library loading during exit - the classLoaderBlocksMutex is sufficient
to ensure that.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>