Fix deadlock during jitdump when a compilation interruption occurs
This change fixes a deadlock which can occur right before the VM is
about to terminate. The following conditions can be observed:

- The jitdump file contains `=== EXCEPTION THROWN (JIT Shutdown) ===`
or similar message
- The crashed thread is generating the jitdump and an interruption
exception is thrown forcing the thread to shut down
- A number of other compilation threads have crashed and are waiting to
acquire the dump lock so they can execute their dump agents

The deadlock happens because the thread generating the jitdump has not
unwound the stack to release the dump lock and so the other crashed
threads can never proceed. The exception that is most commonly thrown
is the `JitShutdown` exception which is thrown in
`acquireVMaccessIfNeeded` when we check if the compilation should be
interrupted. The location where we try to acquire VM access is on the
path which dumps Known Object Table during the jitdump process. Because
we are lacking a try/catch block around this piece of code, any
exception which occurs will flow up and past the dump locking code in
`dmpagent.c`, and all the way past the crash point until we reach the
catch handler in the crashed thread. This means the dump lock never
gets released and we are in a deadlock situation with other crashed
threads just waiting around.

The fix is to simply wrap the jitdump process in a try/catch block and
handle such exceptions by reporting the error and aborting the jitdump
process.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>