Fix deadlock when VM access is requested while freeing the KOT
This change fixes a deadlock which can occur when freeing the Known
Object Table (KOT). The following conditions can be observed:

- The application thread wants to queue an asynchronous compile, but it
can't because it wants to acquire the comp monitor which the diagnostic
thread owns
- The diagnostic thread acquired the comp monitor in
`postCompilationTasks` and is trying to free the known object table,
for which it requires VM access. It tries to get it but it can't
because some other compilation thread has requested exclusive.
- The compilation thread which requested exclusive is trying to reserve
a codecache but it can't get exclusive because the application thread
has VM access since it is still executing

This results in a three way deadlock. The fix here is to ensure in the
locations where we free the KOT we do not hold the comp monitor before
acquiring VM access. To ensure no future bugs occur in this area we
also add a fatal assert in the VM access acquire logic to ensure the
comp monitor is not held.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>