Validate Class Chain of Root Class
Currently, the JIT assumes that if there is a method that can be loaded,
which was acquired using the J9ROMMethod of the J9Method that that VM
requested a compilation of, then the J9Class of the method being
compiled (henceforward referred to as the Root Class) must have the same
layout. However, it is possible for the J9ROMClass of some super class
of the Root Class to be modified on disk, such that the JVM has to load
the new jar file, mark the old J9ROMClass of the super class in the SCC
as stale, and add the new J9ROMClass (if there's space). If this
happens, the J9ROMClass of the Root Class won't change, but the object
whose type is the Root Class could now have its layout changed, in which
case, the AOT code is no longer valid.

This commit stores the class chain of the Root Class as a condition for
eligibility for AOT Compilation, and validates the class chain as a
condition to load the AOT method.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>