Fix static field cycle value tests
There are persistent failures for
testCyclicalStaticFieldDefaultValues{1|2|3} running with -Xjit:count=0.
These were traced back to the calls to Class.newInstance in those tests.
When a particular optimization is enabled, the JIT will end up calling
jitTranslateNewInstanceMethod.  But it looks like the implementation
doesn’t currently handle value types, and we end up with an
InstantiationError being thrown.

The current solution is to avoid calloing Class.newInstance in the tests
until the behaviour of that call is updated.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>