Delegate implementor collection to JITServer client
Alternative implementation of #10118.
Instead of walking the class tree on the server, constructing
the list of resolved implementor methods that are not currently
cached and asking client to provide them, this commit delegates
all work to the client.
Every `collectImplementorsCapped` call is delegated to the client,
which finds and creates all required resolved methods and sends
method infos back to the server, so that it can create server-side
methods.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>