JEP 371 implementation part 3
This change adds code to handle ClassOption.STRONG
1) For hidden class defined without ClassOption.STRONG, 
set its classLoader to anon classLoader. J9_FINDCLASS_FLAG_ANON is 
set for such classes. 
2) For hidden class defined with ClassOption.STRONG, set its 
classLoader to its hostclass's classLoader. J9_FINDCLASS_FLAG_ANON 
is not set for such classes.
3) With 1) and 2) the field that I added in my previous change
ROMClassSegmentAllocationStrategy::__allocNewSeg is not necessary
anymore. Remove this field.
4) Add hidden class in 2) to the classTable of its hostclass's 
ClassLoader. In this way, the current balanced GC code is able to 
find such class when iterating through the classTable. 

Adding hidden classes to the classTable makes it not sharable via 
shared classes cache, as its name won't be unique across JVMs. The 
change in 4) is temporary if we finally decide to change balance 
GC to scan the ram segment like other collectors. A flag can be 
used to control whether to skip hidden class when iterating the 
class table. 

Issue #9328
Fixes #10249

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>