Set Class.classData before init in ClassLoader_defineClassImpl1
Before the new JEP371 changes, the classData could be set in
Access.defineClass in Java code before the class was initialized.

With the new JEP371 changes, all code to define and init the class was
moved into native code, ClassLoader_defineClassImpl1.

This requires Class.classData to be initialized within
ClassLoader_defineClassImpl1 before the class is initialized.

Class.setClassData is removed since it is no longer used.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>