Fix TR_VMFieldsInfo constructor
Previously, the TR_VMFieldInfo constructors assumed that the number of
fields in an object instance must equal the number of slots in the
instance. Value type flattening breaks that assumption, so this
commit refactors the code to correctly handle flattened fields.

The original code traversed all the fields in an object instance and
recorded the number of object references visited. This number was then
used to calculate the size of the dynamically allocated `_gcDescriptor`
array, which was then populated with the offset of each ref slot.
Unfortunately, when enabling value type flattening, the number of ref
slots can end up being larger than the number of ref fields in the object,
resulting writes past the end of the array (which happens to corrupt the
stack because it just happens to get allocated in a stack region).

As a fix, this commit changes `_gcDescriptor` to be a std::vector that
is populated by calling `push_back()` for every ref slot in an instance.
As a result, pre-calculating the number of ref-slots is no longer needed
and the allocated space is guaranteed to be large enough. As a bonus,
the code can be significantly simplified by removing some of the
bookkeeping previously used to calculate the number of ref slots.

In order to make the code easier to understand for myself while fixing
the above issue and for future developers, I have also done the
following refactors:

- Added member initialization list to ensure all members always get
  initialized with some default value (previously, there was a path were
  `_statics` and `_fields` could be uninitialized)
- Moved variable declarations to where the variables are actually used
  (there's no reason to pretend a constructor has to compile with C89)
- Removed unused/unneeded variables and fields:
  - _ramClass
  - _numRefSlotsInObject
  - totalNumFields
  - numSlotsInObject
- Changed return type of `buildField()` to void
  - we no longer need to know if the field that was added is a ref
- Moved field collection to it's own function
  - simplified collection of fields by removing need to care about
    static vs non-static fields
  - we no longer need to know which fields occupy slots in an instance
- Made some variables `const` to make sure they are never assigned
- Removed unused methods:
  - `print()` (only used in commented out code)
  - `getNumRefSlotsInObject()`
- Fixed incorrect indenting

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>