Rework trampolines for POWER10
Since POWER10 supports PC-relative addressing, it is no longer necessary
to use the pTOC in trampolines. Instead, a PC-relative load can be used
to materialize the address to jump to. This also eliminates the need for
temporary trampolines, since trampolines can be atomically patched
in-place.

The trampoline size has also been set to be a fixed size of 32 bytes.
This is done since indirect branch target prediction works based off of
32-byte blocks, meaning that multiple indirect branches in the same
32-byte block will alias each other, causing unnecessary mispredictions.

Co-authored-by: Julian Wang <zlwang@ca.ibm.com>
Signed-off-by: Ben Thomas <ben@benthomas.ca>