Move CPU detection before code cache initialization
With the introduction of POWER10, how the code cache is set up will
depend on whether or not POWER10 support is enabled. However, code cache
initialization is currently being run prior to CPU detection. In order
to make things work correctly, the correct CPU must be detected earlier
in the bootstrap process.

Signed-off-by: Ben Thomas <ben@benthomas.ca>