Allow SDK_FILENAME to be configured via variable file
Give the option to allow SDK_FILENAME to be configurable
via the defaults.yml or any vendor variable file. Set it
based on BUILD_IDENTIFIER and fallback to 'default'.
Default to the existing definition for OpenJ9. Move the sdk
variables setup into the Archive SDK function. We don't
need it set earlier and this allows FULL_SDK_VERSION
to be set from the OpenJDK source tag. Also allow variables
to be used in the SDK_FILENAME template. The Groovy template
engine is used to resolve the variables defined in the yml.

[skip ci]

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>