Improve the mechanism of accepting new request at JITServer
To reduce network communication overhead, the JITServer keeps a cache for
the CHTable (segregated by client UID), which is kept up-to-date with
incremental updates sent by the client. The client keeps track of all the
"new" changes to the CHTable since the last update was sent. When it needs
to send a new compilation request, the client packs all these updates into
a std::string and sends them together with the compilation request.
At the same time, the set of "new" CHTable changes is reset and built anew.

For this mechanism to work, the set of updates needs to be applied at the
server in the same order it was computed at the client. The proper ordering
is accomplished with a sequencing number scheme: when the client computes the
set of CHTable changes to be sent, it will acquire a "ticket" which is nothing
more than a sequenceNumber that is sent together with the updates.
The JITServer receives the updates and must ensure that they are applied in
the order dictated by the sequenceNumber. If an update arrives out-of-order,
the client will "park" the corresponding compilation thread (make it sleep on
a monitor) until its turn comes. The waiting on the monitor has a timeout to
avoid scenarios where a sequenceNumber is completely lost. In that situation,
the JITServer must clear its cache because some important updates may be missing.
Starting with a clear CHTable cache, the JITServer will ask the client for the
entire CHTable, after which the incremental CHTable updates mechanism may start
again.

The mechanism described above is already implemented, but has a shortcoming:
the incoming requests are processed strictly in the order dictated by the client
even though they may not cary any information related to CHTable updates.
This PR relaxes this requirement.

We can distinguish two types of compilation requests: the ones that carry class
load/unload information (critical requests) and the ones that do not.
We must process critical requests strictly in the order in which they were encountered
at the client based on their sequence number. Non-critical requests may be executed
in any order, however, we will impose the following rules:
1) A non-critical request can be executed after a critical request with a higher seqNo.
The reason is that the non-critical request will have more up-to-date information.
2) The opposite is not allowed: a non-critical request with a higher seqNo than a
critical request must wait for the critical request to be processed first.

Implementation
The client will send not only the SeqNo of the current request, but also the SeqNo
of the last critical request. The server receives request seqNo_N and the last critical
request seqNo_M. It then checks whether seqNo_M has been processed and if it did,
then this request can go as well. If not, this request will be parked (put to sleep
into a waiting queue). When a request has been processed, it may need to notify
other requests that are waiting in the queue. It needs to notify all waiting
requests up to and including the next critical request (if there is any).
Only critical requests need to activate threads because a comp request can only
be waiting for a critical request. When activating, we must check that the request
that is being activated does not depend on a critical request that hasn't arrived yet.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>